+++ src/main/java/com/gogym/common/annotation/LoginMemberId.java
@@ -1,4 +1,4 @@
-package com.gogym.aop;
+package com.gogym.common.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
