+++ src/main/java/com/gogym/config/WebConfig.java
@@ -1,14 +1,13 @@
 package com.gogym.config;
 
-import com.gogym.common.annotation.LoginMemberIdHandler;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
+import com.gogym.aop.LoginMemberIdHandler;
 import java.util.List;
 
-// 어노테이션 등록 Config 클래스.
+// 어노테이션을 스프링에 등록하는 클래스
 
 @Configuration
 @RequiredArgsConstructor
